+++ NeoMathEngine/src/CPU/CpuMathEngine.h
@@ -16,6 +16,7 @@ limitations under the License.
 #pragma once
 
 #include <NeoMathEngine/NeoMathEngine.h>
+#include <NeoMathEngine/SimdMathEngine.h>
 #include <RawMemoryManager.h>
 #include <DllLoader.h>
 #include <mutex>
@@ -464,6 +465,7 @@ private:
 
 	CDllLoader dllLoader; // loading library for simd instructions
 	std::unique_ptr<const ISimdMathEngine> simdMathEngine; // interface for using simd instructions
+	SgemmFunc CustomSgemmFunction; // Used when it is availabled and is faster then default sgemm
 
 	IMathEngine& mathEngine() { IMathEngine* engine = this; return *engine; }
 
@@ -496,23 +498,23 @@ private:
 		int matrixHeight, int matrixWidth);
 	void multiplyMatrixByMatrix( const float* firstHandle, int firstHeight,
 		int firstWidth, int firstRowSize, const float* secondHandle, int secondWidth, int secondRowSize,
-		float* resultHandle, int resultRowSize );
+		float* resultHandle, int resultRowSize, SgemmFunc customSgemm = 0 );
 	void multiplyMatrixByMatrixAndAdd( const float* first, int firstHeight,
 		int firstWidth, int firstRowSize, const float* second, int secondWidth, int secondRowSize,
-		float* result, int resultRowSize );
+		float* result, int resultRowSize, SgemmFunc customSgemm = 0 );
 	void multiplyTransposedMatrixByMatrix( const float* first, int firstHeight, int firstWidth,
-		const float* second, int secondWidth, float* result );
+		const float* second, int secondWidth, float* result, SgemmFunc customSgemm = 0 );
 	void batchMultiplyTransposedMatrixByMatrix( int batchSize, const float* first, int firstHeight, int firstWidth, 
 		const float* second, int secondWidth, float* result );
 	void multiplyTransposedMatrixByMatrixAndAdd( const float* first, int firstHeight, int firstWidth, int firstRowSize,
-		const float* second, int secondWidth, int secondRowSize, float* result, int resultRowSize );
+		const float* second, int secondWidth, int secondRowSize, float* result, int resultRowSize, SgemmFunc customSgemm = 0 );
 	void multiplyMatrixByTransposedMatrix( const float* first, int firstHeight,
 		int firstWidth, int firstRowSize, const float* second, int secondHeight, int secondRowSize,
-		float* result, int resultRowSize );
+		float* result, int resultRowSize, SgemmFunc customSgemm = 0 );
 	void batchMultiplyMatrixByTransposedMatrix( int batchSize, const CConstFloatHandle& firstHandle, int firstHeight,
 		int firstWidth, const CConstFloatHandle& secondHandle, int secondHeight, const CFloatHandle& resultHandle );
 	void multiplyMatrixByTransposedMatrixAndAdd( const float* first, int firstHeight, int firstWidth, int firstRowSize,
-		const float* second, int secondHeight, int secondRowSize, float* result, int resultRowSize );
+		const float* second, int secondHeight, int secondRowSize, float* result, int resultRowSize, SgemmFunc customSgemm = 0 );
 
 	template<class T>
 	void blobMergeByDimCommon( int dimNum, const CBlobDesc* from, const CTypedMemoryHandle<T>* fromData, int fromCount,
